+++ source/types/types_str2int.c
@@ -6,17 +6,18 @@
 /*   By: ancoulon <ancoulon@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/04 10:48:29 by ancoulon          #+#    #+#             */
-/*   Updated: 2021/05/27 12:33:31 by ancoulon         ###   ########.fr       */
+/*   Updated: 2021/05/28 13:20:15 by ancoulon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "carbon/types.h"
+#include <stdint.h>
 
-int	types_str2int(char *s)
+int64_t	types_str2int(char *s)
 {
-	int				i;
-	long int		nbr;
-	int				sign;
+	int		i;
+	int64_t	nbr;
+	int		sign;
 
 	i = 0;
 	nbr = 0;
